pyproject.toml
[project]
name = "subtext"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.8.0",
    "numpy>=1.26.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/subtext"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
